[RISCV] Expand codegen -> compression sanity checks and move to a single file
The objdump tests interfere with update_llc_test_checks.py and can't be
automatically update them. Put the sanitify check for compression on the
codegen codepath into a separate file, and expand it to also include tests of
integer materialisation. This would catch changes such as those triggered by 
D41949.


git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@330288 91177308-0d34-0410-b5e6-96231b3b80d8